third_party/WebKit/LayoutTests/fast/canvas-api/canvas-overloads-fillText.html

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<body>
<script>
var ctx = document.createElement('canvas').getContext('2d');

function ExpectedNotEnoughArgumentsMessage(num) {
    return "\"TypeError: Failed to execute 'fillText' on 'CanvasRenderingContext2D': 3 arguments required, but only " + num + " present.\"";
}

test(function(t) {
    assert_throws(null, function() {ctx.fillText();}, ExpectedNotEnoughArgumentsMessage(0));
    assert_throws(null, function() {ctx.fillText('moo');}, ExpectedNotEnoughArgumentsMessage(1));
    assert_throws(null, function() {ctx.fillText('moo',0);}, ExpectedNotEnoughArgumentsMessage(2));
    assert_equals(ctx.fillText('moo',0,0), undefined);
    assert_equals(ctx.fillText('moo',0,0,0), undefined);
    assert_equals(ctx.fillText('moo',0,0,0,0), undefined);    
}, "Test the behavior of CanvasRenderingContext2D.fillText() when called with different numbers of arguments.");
</script>
</body>
